Apply conditions to local value as well as patch
In chartsync we have a local value representing the FluxHelmRelease in
the cluster, and we update its conditions as we attempt to release the
chart as specified. The updates are considered against the current
conditions; so when calculating the changes to conditions, we have to
take into account those already in effect.

As an approximation to checking the cluster value each time, mutate
the local value, so that later changes will be considered against
those earlier.